requirements.txt
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
python-dotenv==1.0.1
shiboken6==6.8.1
